+++ aim/sdk/artifacts/correlation.py
@@ -1,11 +1,12 @@
 from typing import Any
 
-from aim.sdk.artifacts.serializable import Serializable
-from aim.sdk.artifacts.utils import get_pt_tensor
 from aim.engine.utils import get_module
+from aim.sdk.artifacts.artifact import Artifact
+from aim.sdk.artifacts.record import Record
+from aim.sdk.artifacts.utils import get_pt_tensor
 
 
-class Correlation(Serializable):
+class Correlation(Artifact):
     cat = ('correlation',)
 
     def __init__(self, name: str, value: Any, labels):
@@ -18,21 +19,18 @@ class Correlation(Serializable):
     def __str__(self):
         return '{name}'.format(name=self.name)
 
-    def serialize(self) -> dict:
-        corr = self.get_correlation()
-        serialized = {
-            self.LOG_FILE: {
-                'name': self.name,
-                'cat': self.cat,
-                'content': corr,
-                'mode': self.CONTENT_MODE_APPEND,
-                'data': {
-                    'labels': self.labels
-                },
-            },
-        }
+    def serialize(self) -> Record:
+        return Record(
+            name=self.name,
+            cat=self.cat,
+            content=self.get_correlation(),
+            data={
+                'labels': self.labels
+            }
+        )
 
-        return serialized
+    def save_blobs(self, name: str, abs_path: str = None):
+        pass
 
     def get_correlation(self) -> list:
         np = get_module('numpy')